agents/rd-agent-mcp/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rd-agent-mcp"
version = "0.1.0"
description = "Research Agent MCP Server with LangGraph/LangChain/LangSmith"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Scott Weeden", email = "sweeden@ttu.edu"}
]
keywords = ["mcp", "research", "langgraph", "langchain", "agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pypdf>=5.0.0",
    "mcp[cli]>=1.0.0",
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langsmith>=0.1.0",
    "httpx>=0.27.0",
    "chromadb>=0.4.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[project.scripts]
rd-agent-mcp = "rd_agent_mcp.__main__:main"

[project.urls]
Homepage = "https://github.com/sweeden-ttu/rd-agent-mcp"
Repository = "https://github.com/sweeden-ttu/rd-agent-mcp"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
